Build documentation and automation
This is a basic Docker based build project that is intended to be used
on Ubuntu Xenial. The Docker container runs Debian Buster and it builds
three projects:

  The Shift Crypto 2FA Android .apk application
  The Shift Crypto DBBv1 mcu firmware
  The Shift Crypto Bitbox Wallet application

The 2FA application is available as a standard Android package for use
with Android devices. The MCU bootloader and firmware is reproducible
and deterministic within a given Docker container build cycle. The
BitBox Wallet application is available as a Debian package, an RPM
package, and as an AppImage.